requirements.txt
amqp==5.1.1
APScheduler==3.10.1
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coverage==4.0.3
crontab==1.0.1
Django==4.2
django-apscheduler==0.6.2
django-crontab==0.7.1
django-filter==23.1
django-rest-framework==0.1.0
django-restframework==0.0.1
django-rq==2.7.0
djangorestframework==3.14.0
freezegun==1.2.2
kombu==5.2.4
Pillow==9.5.0
prompt-toolkit==3.0.38
pycodestyle==2.10.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
redis==4.5.4
redis-workers==0.1.1
rq==1.13.0
rq-scheduler==0.13.0
six==1.16.0
soft-delete==0.2.2
sqlparse==0.4.3
tzdata==2023.3
tzlocal==4.3
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.6